refactor: выделить класс EventUpdater
